Widen document Revision int → long across the revision metadata column (#3733)
Breaking change for Marten 9.0. The numeric revision tracking column (`mt_version` when
`UseNumericRevisions` is on) is now `bigint` rather than `integer`, and all .NET-side
surfaces that carry the revision widen from `int` to `long`:

  - `Marten.Metadata.IRevisioned.Version`
  - `DocumentMetadata.CurrentRevision`
  - `IRevisionedOperation.Revision`
  - `IDocumentSession.UpdateRevision` / `TryUpdateRevision` parameter
  - `IDocumentStorage<T>.Store(IMartenSession, T, long)`
  - `VersionTracker.RevisionFor` / `StoreRevision` / `RevisionsFor`
  - `MetadataColumn<int>` → `MetadataColumn<long>` for `RevisionColumn`
  - Bulk-loader expected-version argument (`bigint`)
  - `mt_upsert_*` / `mt_update_*` / `mt_overwrite_*` function signatures and local
    variables (`final_version` / `current_version`) now `BIGINT`
  - `MartenRegistry.MetadataConfig.Revision` exposed as `Column<long>`

Schema migration from a Marten 8 deployment is non-destructive: the existing
`integer` column is widened in place via `ALTER COLUMN ... TYPE bigint`, preserving
all stored revision values. Switching from a Guid-versioned column to numeric
revisions still drops and recreates the column.

Also fixes an NRE in `MetadataColumn.Member.set` when the rejected member-type
diagnostic ran on the first (null `_member`) assignment.

User-visible code that implements `IRevisioned` or carries a `[Version]` numeric
property must widen the property to `long`. The migration guide (#4355) covers
this.

+++ src/DaemonTests/Aggregations/side_effects_in_aggregations.cs
@@ -400,7 +400,7 @@ public class SideEffects1: IRevisioned
     public int B { get; set; }
     public int C { get; set; }
     public int D { get; set; }
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 public record WasDeleted(Guid Id);
@@ -460,7 +460,7 @@ public class SideEffects2: IRevisioned
     public int B { get; set; }
     public int C { get; set; }
     public int D { get; set; }
-    public int Version { get; set; }
+    public long Version { get; set; }
 }
 
 public class RecordingMessageOutbox: IMessageOutbox
